app/requirements.txt
fastapi==0.85.0
uvicorn==0.18.2
sqlalchemy==1.4.32
pydantic==1.10.2
jinja2==3.1.2
openai==0.28.0
python-dotenv==0.20.0
requests==2.28.1
gunicorn==20.1.0
alembic==1.13.1
psycopg2